bug fix edge case that would lead to panic

+++ service/vm_service/internal_service/cs_service.go
@@ -4,7 +4,6 @@ import (
 	"fmt"
 	vmModel "go-deploy/models/vm"
 	"go-deploy/pkg/conf"
-	"go-deploy/pkg/status_codes"
 	"go-deploy/pkg/subsystems/cs"
 	"go-deploy/pkg/subsystems/cs/commands"
 	csModels "go-deploy/pkg/subsystems/cs/models"
@@ -48,6 +47,11 @@ func CreateCS(name, sshPublicKey string) (*CsCreated, error) {
 	vm, err := vmModel.GetByName(name)
 	if err != nil {
 		return nil, makeError(err)
+	}
+
+	if vm == nil {
+		// if vm does not exist, don't treat as error, don't create -> job will not fail
+		return nil, nil
 	}
 
 	// vm
@@ -193,6 +197,10 @@ func DeleteCS(name string) error {
 		return makeError(err)
 	}
 
+	if vm == nil {
+		return nil
+	}
+
 	if vm.Subsystems.CS.PortForwardingRule.ID != "" {
 		err = client.DeletePortForwardingRule(vm.Subsystems.CS.PortForwardingRule.ID)
 		if err != nil {
@@ -230,58 +238,6 @@ func DeleteCS(name string) error {
 	}
 
 	return nil
-}
-
-func GetStatusCS(name string) (int, string, error) {
-	makeError := func(err error) error {
-		return fmt.Errorf("failed to get status for cs vm %s. details: %s", name, err)
-	}
-
-	unknownMsg := status_codes.GetMsg(status_codes.ResourceUnknown)
-
-	client, err := withClient()
-	if err != nil {
-		return status_codes.ResourceUnknown, unknownMsg, makeError(err)
-	}
-
-	vm, err := vmModel.GetByName(name)
-	if err != nil {
-		return status_codes.ResourceUnknown, unknownMsg, makeError(err)
-	}
-
-	csVmID := vm.Subsystems.CS.VM.ID
-	if csVmID == "" {
-		return status_codes.ResourceNotFound, status_codes.GetMsg(status_codes.ResourceNotFound), nil
-	}
-
-	status, err := client.GetVmStatus(csVmID)
-	if err != nil {
-		return status_codes.ResourceUnknown, unknownMsg, makeError(err)
-	}
-
-	var statusCode int
-	switch status {
-	case "Starting":
-		statusCode = status_codes.ResourceUnknown
-	case "Running":
-		statusCode = status_codes.ResourceRunning
-	case "Stopping":
-		statusCode = status_codes.ResourceStopping
-	case "Stopped":
-		statusCode = status_codes.ResourceStopped
-	case "Migrating":
-		statusCode = status_codes.ResourceRunning
-	case "Error":
-		statusCode = status_codes.ResourceError
-	case "Unknown":
-		statusCode = status_codes.ResourceUnknown
-	case "Shutdowned":
-		statusCode = status_codes.ResourceStopped
-	default:
-		statusCode = status_codes.ResourceUnknown
-	}
-
-	return statusCode, status_codes.GetMsg(statusCode), nil
 }
 
 func AttachGPU(gpuID, vmID string) error {
